+++ src/main/java/net/ccbluex/liquidbounce/event/async/Extensions.kt
@@ -14,48 +14,8 @@ import net.ccbluex.liquidbounce.event.EventHook
 import net.ccbluex.liquidbounce.event.EventManager
 import net.ccbluex.liquidbounce.event.Listenable
 import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 
-
-/**
- * Wait until given [condition] returns true.
- *
- * @param condition It will be called on [Dispatchers.Main] (the Render thread)
- * @return Total ticks during waiting
- */
-suspend inline fun waitUntil(crossinline condition: () -> Boolean): Int =
-    suspendCoroutine { cont ->
-        var waitingTick = 0
-        TickScheduler.schedule {
-            waitingTick++
-            try {
-                if (condition()) {
-                    cont.resume(waitingTick)
-                    true
-                } else {
-                    false
-                }
-            } catch (e: Throwable) {
-                cont.resumeWithException(e)
-                true
-            }
-        }
-    }
-
-/**
- * Wait for given [ticks].
- */
-suspend fun waitTicks(ticks: Int) {
-    require(ticks >= 0) { "Negative tick: $ticks" }
-
-    if (ticks == 0) {
-        return
-    }
-
-    var remainingTick = ticks
-    waitUntil { --remainingTick == 0 }
-}
 
 /**
  * Take next event instance of given type and run [handler] with it.
@@ -75,10 +35,11 @@ inline fun <reified E : Event> Listenable.takeNext(
  *
  * @return the event instance
  */
-suspend inline fun <reified E : Event> Listenable.waitNext(priority: Byte = 0): E =
-    suspendCoroutine { cont ->
-        takeNext(priority, cont::resume)
-    }
+suspend inline fun <reified E : Event> Listenable.waitNext(
+    priority: Byte = 0
+): E = suspendCoroutine { cont ->
+    takeNext(priority, cont::resume)
+}
 
 /**
  * Wait next event instance of given type which matches [predicate].
@@ -96,26 +57,6 @@ suspend inline fun <reified E : Event> Listenable.waitNext(
             return next
         }
     }
-}
-
-/**
- * Waits until the fixed amount of [ticks] ran out or the [callback] returns true.
- */
-suspend inline fun waitConditional(
-    ticks: Int,
-    crossinline callback: (elapsedTicks: Int) -> Boolean
-): Boolean {
-    require(ticks >= 0) { "Negative tick: $ticks" }
-
-    if (ticks == 0) {
-        return true
-    }
-
-    var elapsedTicks = 0
-    // `elapsedTicks` in 0 until `ticks`
-    waitUntil { elapsedTicks >= ticks || callback(elapsedTicks++) }
-
-    return elapsedTicks >= ticks
 }
 
 /**
